# [Bug] Fix rocshmem_p and rocshmem_g functional tests